tick by curve and hero speed
delay = 1/(speed + valueOfCurve)

+++ Assets/Scripts/SOScripts/HeroSO.cs
@@ -15,7 +15,7 @@ public class HeroSO : ScriptableObject
 {
     public string nameOfHero;
     public int health;
-    public int speed;//vitesse ?
+    public float speed;//vitesse ?
     public HeroesInfo classe;
     public VisionType visionType;
     public Aggressivity aggressivity;
